activity_main.xml change about us page
