+++ app/src/main/AndroidManifest.xml
@@ -9,9 +9,9 @@
         android:roundIcon="@mipmap/ic_launcher_round"
         android:supportsRtl="true"
         android:theme="@style/AppTheme">
-        <activity android:name=".Aboutus"></activity>
-        <activity android:name=".contactus"></activity>
-        <activity android:name=".splech_screen" />
+        <activity android:name=".login" />
+        <activity android:name=".Aboutus" />
+        <activity android:name=".contactus" />
         <activity android:name=".MainActivity">
             <intent-filter>
                 <action android:name="android.intent.action.MAIN" />
