+++ app/src/main/java/com/example/mygroup/Aboutus.java
@@ -1,14 +1,15 @@
 package com.example.mygroup;
 
-import androidx.appcompat.app.AppCompatActivity;
-
 import android.os.Bundle;
 
-public class Aboutus extends AppCompatActivity {
+import androidx.appcompat.app.AppCompatActivity;
 
+public class Aboutus extends AppCompatActivity {
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_aboutus);
+
     }
+
 }
