+++ app/src/main/java/com/example/mygroup/MainActivity.java
@@ -1,78 +1,24 @@
 package com.example.mygroup;
 
-import androidx.appcompat.app.AppCompatActivity;
-
+import android.app.Activity;
 import android.content.Intent;
 import android.os.Bundle;
 import android.os.Handler;
-import android.view.View;
-import android.widget.Button;
-import android.widget.EditText;
-import android.widget.Toast;
 
-public class MainActivity extends AppCompatActivity {
-
-    /* Define the UI elements */
-    private EditText eName;
-    private EditText ePassword;
-
-    String userName = "";
-    String userPassword = "";
-
-    /* Class to hold credentials */
-    static class Credentials {
-        String name = "Admin";
-        int password = 123456;
-    }
-
-    boolean isValid = false;
+public class MainActivity extends Activity {
 
     @Override
     protected void onCreate(Bundle savedInstanceState) {
         super.onCreate(savedInstanceState);
         setContentView(R.layout.activity_main);
-
-        /* Bind the XML views to Java Code Elements */
-        eName = findViewById(R.id.username);
-        ePassword = findViewById(R.id.password);
-        Button eLogin = findViewById(R.id.go);
-
-        /* Describe the logic when the login button is clicked */
-        eLogin.setOnClickListener(new View.OnClickListener() {
+        int SPLASH_TIME_OUT = 4000;
+        new Handler().postDelayed(new Runnable() {
             @Override
-            public void onClick(View view) {
-
-                /* Obtain user inputs */
-                userName = eName.getText().toString();
-                userPassword = ePassword.getText().toString();
-
-                /* Check if the user inputs are empty */
-                if (userName.isEmpty() || userPassword.isEmpty()) {
-                    /* Display a message toast to user to enter the details */
-                    Toast.makeText(MainActivity.this, "Please enter name and password!", Toast.LENGTH_LONG).show();
-
-                } else {
-
-                    /* Validate the user inputs */
-                    isValid = validate(userName, userPassword);
-                    startActivity(new Intent(MainActivity.this, contactus.class));
-                    /* Validate the user inputs */
-
-                    /* If not valid */
-
-
-                }
+            public void run() {
+                Intent mainIntent = new Intent(MainActivity.this, login.class);
+                startActivity(mainIntent);
+                finish();
             }
-        });
+        }, SPLASH_TIME_OUT);
     }
-
-    /* Validate the credentials */
-    private boolean validate(String userName, String userPassword) {
-        /* Get the object of Credentials class */
-        Credentials credentials = new Credentials();
-
-        /* Check the credentials */
-        return userName.equals(credentials.name) && userPassword.equals(credentials.password);
-    }
-
 }
