+++ app/src/main/res/drawable/lg.PNG
