+++ app/src/main/res/drawable/roundened_corner.xml
@@ -1,5 +1,5 @@
 <?xml version="1.0" encoding="utf-8"?>
 <shape xmlns:android="http://schemas.android.com/apk/res/android">
     <corners android:radius="24dp" />
-    <solid android:color="#E91E63" />
+    <solid android:color="#1e88e5" />
 </shape>
