--- app/src/main/res/layout/activity_splech_screen.xml
@@ -1,31 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-
-<RelativeLayout xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".MainActivity"
-    android:background="@drawable/background">
-
-
-<androidx.constraintlayout.widget.ConstraintLayout
-
-
-    android:layout_width="match_parent"
-    android:layout_height="match_parent"
-    tools:context=".splech_screen">
-
-    <ImageView
-        android:id="@+id/imageView3"
-        android:layout_width="183dp"
-        android:layout_height="202dp"
-        app:layout_constraintBottom_toBottomOf="parent"
-        app:layout_constraintEnd_toEndOf="parent"
-        app:layout_constraintStart_toStartOf="parent"
-        app:layout_constraintTop_toTopOf="parent"
-        app:srcCompat="@drawable/logo"
-        android:contentDescription="@string/todo" />
-
-</androidx.constraintlayout.widget.ConstraintLayout>
-</RelativeLayout>
