+++ app/src/main/res/values/colors.xml
@@ -3,6 +3,6 @@
     <color name="colorPrimary">#6200EE</color>
     <color name="colorPrimaryDark">#3700B3</color>
     <color name="colorAccent">#03DAC5</color>
-    <color name="colorfont">#E91E63</color>
+    <color name="colorfont">#1e88e5</color>
     <color name="colorback">#8BC34A</color>
 </resources>
