+++ app/src/main/res/values/styles.xml
@@ -6,5 +6,12 @@
         <item name="colorPrimaryDark">@color/colorPrimaryDark</item>
         <item name="colorAccent">@color/colorAccent</item>
     </style>
+    <style name="AppTheme.NoTitle">
+            <item name="android:windowNoTitle">true</item>
+         
+            <!-- Support library compatibility -->
+            <item name="windowActionBar">false</item>
+    </style>
+                    
 
 </resources>
